# api/openapi.yaml
openapi: 3.1.0
info:
  title: Bramble Mesh Node API
  description: >
    JSON-RPC 2.0 API for Bramble mesh nodes.
    Methods are invoked via JSON-RPC over serial (UART), WebSocket, or BLE.
    Each method is mapped to a POST path for codegen and documentation purposes.
    The actual wire format is standard JSON-RPC 2.0.
    WebSocket and BLE connections require the per-device auth token by default;
    serial is unauthenticated by design (physical access is the trust anchor).
    Unauthenticated WS/BLE clients may call only bramble.ping and
    bramble.getVersion; every other method answers Unauthorized (-1005).
  version: 0.7.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://bramble.local
    description: Bramble node HTTP host (WebSocket JSON-RPC upgrades on /ws)
security:
  - deviceToken: []
paths:
...
  /rpc/bramble.exportTopology:
    post:
      operationId: exportTopology
      summary: Export this node's observed mesh state
      description: >
        Returns one node's view of the mesh as a single document: its identity,
        the neighbours it hears with per-link RSSI and SNR, its routing table,
        and the PHY and frequency-plan parameters that decide time-on-air. The
        neighbours and routes arrays are the same shapes bramble.getNeighbors
        and bramble.getRoutes return, written by the same emitters. Collect one
        export per node and feed the files to the simulator's digital-twin
        importer (docs/digital-twin.md) to reconstruct the deployment as a
        runnable scenario.
      tags: [query]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmptyParams'
      responses:
        '400':
          description: Bad request (invalid params or request body)
        '200':
          description: Observed topology export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopologyExportResponse'
...
components:
  schemas:
...
    EmptyParams:
      type: object
      description: No parameters required.
      properties: {}
...
    Neighbor:
      type: object
      description: A directly heard radio neighbor.
      required: [address, rssi, snr, deliveryRate, airtimeRemaining, last_seen_ms]
      properties:
        address:
          type: string
          description: Neighbor address as 8-char uppercase hex.
        rssi:
          type: integer
          description: Received signal strength (dBm).
        snr:
          type: number
          description: Signal-to-noise ratio (dB).
        deliveryRate:
          type: integer
          description: Link delivery rate metric (0-255).
        airtimeRemaining:
          type: integer
          description: Neighbor-advertised airtime budget remaining.
        last_seen_ms:
          type: integer
          description: Milliseconds since last packet received.
        name:
          type: string
          description: Neighbor display name (present only when known).
    Route:
      type: object
      description: A routing table entry.
      required: [dest, next_hop, hop_count, metric, state, use_count]
      properties:
        dest:
          type: string
          description: Destination address as 8-char uppercase hex.
        next_hop:
          type: string
          description: Next-hop address as 8-char uppercase hex.
        hop_count:
          type: integer
          description: Number of hops to destination.
        metric:
          type: integer
          description: Route cost metric (lower is better).
        state:
          type: string
          enum: [discovering, unverified, active, stale, broken, unknown]
          description: Route state.
        use_count:
          type: integer
          description: Number of times this route has been used.
...
    TopologyExportResponse:
      type: object
      description: >
        One node's observed mesh state, shaped for the simulator's digital-twin
        importer. Observation only: every field is state the node already keeps,
        read at the moment of the call.
      required: [twin_schema, node, radio, neighbors, routes]
      properties:
        twin_schema:
          type: integer
          description: >
            Document schema version. The importer refuses a version it does not
            know rather than guessing at fields.
        node:
          type: object
          required: [address, firmware_version, protocol_version, hardware, uptime_s]
          properties:
            address:
              type: string
              description: This node's address as 8-char uppercase hex.
            name:
              type: string
              description: Configured node name (omitted when unset).
            firmware_version:
              type: string
            protocol_version:
              type: string
            hardware:
              type: string
            uptime_s:
              type: integer
              description: Seconds since boot, the age bound on every observation below.
        radio:
          type: object
          description: >
            Runtime PHY plus the compiled-in frequency plan. sf, bw_hz and
            coding_rate are what price a frame's time-on-air; the plan's duty
            cycle bounds what the deployment may spend.
          required:
            - frequency_mhz
            - sf
            - bw_hz
            - coding_rate
            - tx_power_dbm
            - region
            - regulatory
            - max_duty_cycle_pct
            - duty_cycle_enforced
          properties:
            frequency_mhz:
              type: number
            sf:
              type: integer
            bw_hz:
              type: integer
            coding_rate:
              type: integer
              description: 1 to 4, meaning 4/5 through 4/8.
            tx_power_dbm:
              type: integer
            region:
              type: string
              description: Frequency plan name, for example US915.
            regulatory:
              type: string
              description: Regulatory regime the plan follows.
            max_duty_cycle_pct:
              type: integer
              description: Plan duty-cycle ceiling; 100 means no limit.
            duty_cycle_enforced:
              type: boolean
              description: Whether the ceiling is hard-enforced or advisory.
        neighbors:
          type: array
          description: Nodes heard directly, with the link quality they were heard at.
          items:
            $ref: '#/components/schemas/Neighbor'
        routes:
          type: array
          items:
            $ref: '#/components/schemas/Route'
...
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
      description: >
        Per-device auth token (generated on first boot, NVS-persisted).
        WebSocket clients send "Authorization: Bearer <token>" or the ?token=
        query parameter; BLE clients write the token as the first write on a
        new connection; serial is unauthenticated by design. Without the token
        only bramble.ping and bramble.getVersion are callable.

# ── Notifications (extension) ──────────────────────────────────────────────
...